+++ services/networkService.ts
