+++ services/transactionService.ts
@@ -4,7 +4,7 @@ import { Prisma, Transaction, Address } from '@prisma/client'
 import grpcService from 'services/grpcService'
 import { parseAddress } from 'utils/validators'
 import { satoshisToUnit } from 'utils/index'
-import { fetchAddressBySubstring } from 'services/addressesService'
+import { fetchAddressBySubstring } from 'services/addressService'
 import _ from 'lodash'
 import { RESPONSE_MESSAGES, FETCH_N, FETCH_DELAY } from 'constants/index'
 import xecaddr from 'xecaddrjs'
